requirements.txt
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-storages==1.13.1
dj-rest-auth==2.2.5
django-allauth==0.51.0
drf-yasg==1.21.4
psycopg2-binary==2.9.5
pytest-django==4.5.2
pillow==9.3.0
boto3==1.26.20
pymysql==1.0.2